requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster @ file:///home/conda/feedstock_root/build_artifacts/alabaster_1733750398730/work
annotated-types==0.7.0
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1742243108/work
arrow @ file:///home/conda/feedstock_root/build_artifacts/arrow_1733584251875/work
arxiv @ file:///home/conda/feedstock_root/build_artifacts/arxiv_1744118931307/work
asttokens==3.0.0
attrs==19.1.0
Authlib==1.6.0
babel @ file:///home/conda/feedstock_root/build_artifacts/babel_1738490167835/work
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1744783198182/work
bibtexparser @ file:///home/conda/feedstock_root/build_artifacts/bibtexparser_1734879298321/work
biopython @ file:///D:/bld/biopython_1737241762116/work
biorxiv_retriever==0.20.1
Brotli @ file:///D:/bld/brotli-split_1749229870877/work
certifi==2025.4.26
cffi @ file:///D:/bld/cffi_1725560655177/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.2.1
clinical_trials==1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated @ file:///home/conda/feedstock_root/build_artifacts/deprecated_1737986966356/work
distro==1.9.0
docutils @ file:///home/conda/feedstock_root/build_artifacts/docutils_1733217766141/work
dotenv==0.9.9
et_xmlfile @ file:///home/conda/feedstock_root/build_artifacts/et_xmlfile_1733749653422/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing==2.2.0
fake-useragent @ file:///D:/bld/fake-useragent_1744653914642/work
fastmcp==2.9.0
feedparser @ file:///home/conda/feedstock_root/build_artifacts/feedparser_1734808041952/work
fonttools==4.58.0
free_proxy @ file:///home/conda/feedstock_root/build_artifacts/free-proxy_1735329743671/work
frozenlist==1.6.0
greenlet==3.2.2
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1745526374115/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib==1.1
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
httpx-sse==0.4.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imagesize @ file:///home/conda/feedstock_root/build_artifacts/imagesize_1656939531508/work
iniconfig==2.1.0
inquirerpy==0.3.4
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.10.0
json_repair==0.45.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-experimental==0.3.4
langchain-mcp-adapters==0.1.1
langchain-openai==0.3.17
langchain-tavily==0.2.4
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.70
langsmith==0.3.42
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe @ file:///D:/bld/markupsafe_1733219601796/work
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
maxminddb==2.7.0
mcp==1.9.4
mdurl==0.1.2
multidict==6.4.4
mypy==1.16.1
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.81.0
openapi-pydantic==0.5.1
openpyxl @ file:///D:/bld/openpyxl_1725460918321/work
orjson==3.10.18
ormsgpack==1.9.1
outcome @ file:///home/conda/feedstock_root/build_artifacts/outcome_1733406188332/work
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pfzy==0.3.4
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
proxybroker==0.3.2
pure_eval==0.2.3
pycares==4.9.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySocks @ file:///D:/bld/pysocks_1733217287171/work
pytest==8.3.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1742948348/work
python-multipart==0.0.20
pytrials==1.0.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.2
readabilipy==0.3.0
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-toolbelt==1.0.0
rich==14.0.0
roman-numerals-py @ file:///home/conda/feedstock_root/build_artifacts/roman-numerals-py_1740240236123/work
scholarly @ file:///home/conda/feedstock_root/build_artifacts/scholarly_1673917828640/work
selenium @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_selenium_1748305242/work
setuptools==80.8.0
sgmllib3k @ file:///home/conda/feedstock_root/build_artifacts/sgmllib3k_1734543040599/work
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
snowballstemmer @ file:///home/conda/feedstock_root/build_artifacts/snowballstemmer_1747749415787/work
sortedcontainers @ file:///home/conda/feedstock_root/build_artifacts/sortedcontainers_1738440353519/work
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1746563585861/work
Sphinx @ file:///home/conda/feedstock_root/build_artifacts/sphinx_1740956487887/work
sphinx_rtd_theme @ file:///home/conda/feedstock_root/build_artifacts/sphinx_rtd_theme_1730015256242/work
sphinxcontrib-applehelp @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-applehelp_1733754101641/work
sphinxcontrib-devhelp @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-devhelp_1733754113405/work
sphinxcontrib-htmlhelp @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-htmlhelp_1733754280555/work
sphinxcontrib-jquery @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-jquery_1734344508263/work
sphinxcontrib-jsmath @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-jsmath_1733753744933/work
sphinxcontrib-qthelp @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-qthelp_1733753408017/work
sphinxcontrib-serializinghtml @ file:///home/conda/feedstock_root/build_artifacts/sphinxcontrib-serializinghtml_1733750479108/work
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
trio @ file:///D:/bld/trio_1745237152806/work
trio-websocket @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_trio-websocket_1740493956/work
typer==0.16.0
types-python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/types-python-dateutil_1747417193651/work
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
uvicorn==0.34.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client @ file:///home/conda/feedstock_root/build_artifacts/websocket-client_1733157342724/work
wheel==0.45.1
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
wrapt @ file:///D:/bld/wrapt_1736869459848/work
wsproto @ file:///home/conda/feedstock_root/build_artifacts/wsproto_1733060193308/work
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
